# docs/swagger.yaml
basePath: /swdp/v1/api
definitions:
  projects.ProjectResponse:
    properties:
      owner_id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
      project_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      project_name:
        example: SWDP Backend
        type: string
    type: object
  users.User:
    type: object
host: localhost:8282
info:
  contact:
    email: eselasenobleson@gmail.com
    name: Noble Eselase Vulley
  description: Web-Based Software Development Platform Backend
  license:
    name: Proprietary
    url: https://example.com/license
  termsOfService: https://example.com/terms/
  title: SWDP Backend API
  version: "1.0"
paths:
  /projects/all:
    get:
      consumes:
      - application/json
      description: Get all projects for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/projects.ProjectResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List projects
      tags:
      - Projects
  /users/add:
    post:
      consumes:
      - application/json
      description: Creates a new user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/users.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/users.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - User
  /users/all:
    get:
      description: Retrieves all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/users.User'
            type: array
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - User
  /users/delete/{id}:
    delete:
      description: Soft deletes a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - User
  /users/get/{id}:
    get:
      description: Retrieves a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user
      tags:
      - User
  /users/update/{id}:
    put:
      consumes:
      - application/json
      description: Updates user fields
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - User
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"